+++ app/src/androidTest/java/com/github/steroidteam/todolist/CustomMatchers.java
@@ -1,9 +1,12 @@
 package com.github.steroidteam.todolist;
 
 import android.view.View;
+import android.widget.CheckBox;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+import androidx.test.espresso.UiController;
+import androidx.test.espresso.ViewAction;
 import androidx.test.espresso.matcher.BoundedMatcher;
 import java.util.Objects;
 import org.hamcrest.Description;
@@ -39,6 +42,28 @@ public class CustomMatchers {
         };
     }
 
+    public static Matcher<View> atPositionCheckBox(
+            final int position, @NonNull final boolean expectedBox, @NonNull final int layout_id) {
+        return new BoundedMatcher<View, RecyclerView>(RecyclerView.class) {
+            @Override
+            public void describeTo(Description description) {
+                description.appendText(
+                        "View holder at position "
+                                + String.valueOf(position)
+                                + ", expected: "
+                                + expectedBox
+                                + " ");
+            }
+
+            @Override
+            protected boolean matchesSafely(final RecyclerView view) {
+                View taskView = view.getChildAt(position);
+                CheckBox boxView = taskView.findViewById(layout_id);
+                return boxView.isChecked() == expectedBox;
+            }
+        };
+    }
+
     /** Helper to check the size of a recyclerView */
     public static Matcher<View> ItemCountIs(@NonNull final int expectedCount) {
         return new BoundedMatcher<View, RecyclerView>(RecyclerView.class) {
@@ -50,6 +75,28 @@ public class CustomMatchers {
             @Override
             protected boolean matchesSafely(final RecyclerView view) {
                 return Objects.requireNonNull(view.getAdapter()).getItemCount() == expectedCount;
+            }
+        };
+    }
+
+    // Simple ViewAction to click on the button within a item of the recyclerView
+
+    public static ViewAction clickChildViewWithId(final int id) {
+        return new ViewAction() {
+            @Override
+            public Matcher<View> getConstraints() {
+                return null;
+            }
+
+            @Override
+            public String getDescription() {
+                return "Click on a child view with specified id.";
+            }
+
+            @Override
+            public void perform(UiController uiController, View view) {
+                View v = view.findViewById(id);
+                v.performClick();
             }
         };
     }
